[RIA] Added Servlet to respond to Visited Auctions requests
This Servlet reads the list of IDs from the Request and returns a JSON containing the Auction's data from the Auctions whose IDs match the received list.

It does not save or retrieve visited Auctions from the DB or the Backend in any way, they must come from the client through the request.